+++ frontend/src/components/placementVisuals/AddPlacementContainer.tsx
@@ -1,177 +1,177 @@
 import React, { useCallback, useState } from "react"
-import { Alert, Button, Form, Select, Tabs, notification, Row } from "antd"
+import { Button, Select, Tabs, notification, Row } from "antd"
 import { useAppDispatch, useAppSelector } from "../../hooks"
 import { selectCoordinatesByID, selectContainerKindsByID } from "../../selectors"
-import { PLACED_STRING, containerSample } from "./PlacementTab"
 import Modal from "antd/lib/modal/Modal"
 import SearchContainer from "../SearchContainer"
 import Input from "antd/lib/input/Input"
 import api from "../../utils/api"
 import { MAX_CONTAINER_BARCODE_LENGTH, MAX_CONTAINER_NAME_LENGTH, barcodeRules, nameRules } from "../../constants"
+import { FMSContainer, FMSId, FMSSample } from "../../models/fms_api_models"
 
-interface AddPlacementContainerProps {
-    onConfirm: (destinationContainer) => void
-    destinationContainerList: containerSample[]
-    setDestinationIndex: (number: number) => void
+export interface DestinationContainer {
+    container_barcode: string
+    container_name: string
+    container_kind: string
+    samples: { [key in FMSId]: { id: FMSId, coordinates: string, name: string } }
+}
+
+export interface AddPlacementContainerProps {
+    onConfirm: (destinationContainer: DestinationContainer) => void
+    existingContainers: DestinationContainer[]
 }
 
 //component used to handle the creation of a new destination container
-const AddPlacementContainer = ({ onConfirm, destinationContainerList, setDestinationIndex }: AddPlacementContainerProps) => {
+const AddPlacementContainer = ({ onConfirm, existingContainers }: AddPlacementContainerProps) => {
     const [isPopup, setIsPopup] = useState<boolean>(false)
 
     //used to hold loaded container data
-    const [loadedContainer, setLoadedContainer] = useState<any>({})
+    const [loadedContainer, setLoadedContainer] = useState<Partial<DestinationContainer>>({})
     const [selectedTab, setSelectedTab] = useState<string>('new')
-    const [error, setError] = useState<string | undefined>(undefined)
-    const [newContainer, setNewContainer] = useState<any>({})
+    
+    // make at least empty samples required for newly created container
+    const [newContainer, setNewContainer] = useState<Pick<DestinationContainer, 'samples'> & Partial<DestinationContainer>>({ samples: {} })
 
     const coordinates = useAppSelector(selectCoordinatesByID)
     const containerKinds = useAppSelector(selectContainerKindsByID)
 
     const getContainerKindOptions = useCallback(() => {
-      if (containerKinds) {
-        const options = Object.keys(containerKinds).filter(id => containerKinds[id].children_ids.length === 0 && id !== "tube").sort().map((containerKind) => {
-          return {
-            label: containerKind,
-            value: containerKind
-          }
-        })
-        return options
-      }
-      return []
+        if (containerKinds) {
+            const options = Object.keys(containerKinds).filter(id => containerKinds[id].children_ids.length === 0 && id !== "tube").sort().map((containerKind) => {
+                return {
+                    label: containerKind,
+                    value: containerKind
+                }
+            })
+            return options
+        }
+        return []
     }, [containerKinds])
 
     const dispatch = useAppDispatch()
     //retrieves container on search
-    const handleContainerLoad = useCallback(async (loadedContainer) => {
-        const container = await dispatch(api.containers.get(loadedContainer))
-        const containerName = container.data.name
-        const containerBarcode = container.data.barcode
-        const newDestination = {}
-        if (container.data.samples.length > 0) {
-            const loadedSamples = await dispatch(api.samples.list({ id__in: container.data.samples.join(','), limit: 100000}))
+    const handleContainerLoad = useCallback(async (loadedContainer: FMSId) => {
+        const container: FMSContainer = (await dispatch(api.containers.get(loadedContainer))).data
+        const containerName = container.name
+        const containerBarcode = container.barcode
+        const newDestination: DestinationContainer['samples']  = {}
+        if (container.samples.length > 0) {
+            const loadedSamples: FMSSample[] = (await dispatch(api.samples.list({ id__in: container.samples.join(','), limit: 100000 }))).data.results
 
-            loadedSamples.data.results.forEach(sample => {
-                    newDestination[sample.id] = { id: sample.id, coordinates: (sample.coordinate && coordinates[sample.coordinate].name), type: PLACED_STRING, name: sample.name, sourceContainer: containerName }
+            loadedSamples.forEach(sample => {
+                newDestination[sample.id] = { id: sample.id, coordinates: (sample.coordinate && coordinates[sample.coordinate].name) as string, name: sample.name }
             })
         }
-        setLoadedContainer({ container_barcode: containerBarcode, container_name: containerName, container_kind: container.data.kind, samples: { ...newDestination }, })
-    }, [coordinates])
-
-    const containerAlreadyExists = async (container, destinationContainerList: containerSample[]) => {
-      /* Centralized error notifications */
-      const barcodeExistsError = () => {
-        setError("Existing container barcode.")
-        const EXISTING_BARCODE_NOTIFICATION_KEY = `LabworkStep.placement-existing-container-barcode`
-        notification.error({
-          message: `New container barcode already exists.`,
-          key: EXISTING_BARCODE_NOTIFICATION_KEY,
-          duration: 20
-        })
-      }
-      const nameExistsError = () => {
-        setError("Existing container name.")
-        const EXISTING_NAME_NOTIFICATION_KEY = `LabworkStep.placement-existing-container-name`
-        notification.error({
-          message: `New container name already exists.`,
-          key: EXISTING_NAME_NOTIFICATION_KEY,
-          duration: 20
-        })
-      }
-
-      let exists: boolean = false
-      /* Get locally created containers first */
-      /* Check if the barcode is used in the destinationContainerList */
-      let barcode_exists: boolean = destinationContainerList.some(containerSample => containerSample.container_barcode == container.container_barcode)
-      if (barcode_exists) {
-        barcodeExistsError()
-      }
-      /* Check if the name is used in the destinationContainerList */
-      let name_exists: boolean = destinationContainerList.some(containerSample => containerSample.container_name == container.container_name)
-      if (name_exists) {
-        nameExistsError()
-      }
-      exists = name_exists || barcode_exists
-      if (exists) {
-        /* Fast return if already exists in frontend display */
-        return exists
-      }
-
-      /* Get existing containers from freezeman */
-      /* Check if there is already a container in Freezeman with that abrcode */
-      const barcodeResult = await dispatch(api.containers.list({barcode: container.container_barcode}))
-      barcode_exists = barcodeResult.data.count != 0
-      if (barcode_exists) {
-        barcodeExistsError()
-      }
-      /* Check if there is already a container in Freezeman with that name */
-      const nameResult = await dispatch(api.containers.list({name: container.container_name}))
-      name_exists = nameResult.data.count != 0
-      if (name_exists) {
-        nameExistsError()
-      }
-      exists = name_exists || barcode_exists
-      return exists
-    }
+        setLoadedContainer({ container_barcode: containerBarcode, container_name: containerName, container_kind: container.kind, samples: { ...newDestination }, })
+    }, [coordinates, dispatch])
 
     //calls addDestination prop with 'New Destination' container
     const handleConfirm = useCallback(() => {
-        const addContainer = (container) => {
-          onConfirm(container)
-          setIsPopup(false)
-          setNewContainer({})
-          setDestinationIndex(destinationContainerList.length)
+        const containerAlreadyExists = async (container: DestinationContainer, destinationContainerList: DestinationContainer[]) => {
+            /* Centralized error notifications */
+            const barcodeExistsError = () => {
+                const EXISTING_BARCODE_NOTIFICATION_KEY = `LabworkStep.placement-existing-container-barcode`
+                notification.error({
+                    message: `New container barcode already exists.`,
+                    key: EXISTING_BARCODE_NOTIFICATION_KEY,
+                    duration: 20
+                })
+            }
+            const nameExistsError = () => {
+                const EXISTING_NAME_NOTIFICATION_KEY = `LabworkStep.placement-existing-container-name`
+                notification.error({
+                    message: `New container name already exists.`,
+                    key: EXISTING_NAME_NOTIFICATION_KEY,
+                    duration: 20
+                })
+            }
+    
+            let exists = false
+            /* Get locally created containers first */
+            /* Check if the barcode is used in the destinationContainerList */
+            let barcode_exists: boolean = destinationContainerList.some(containerSample => containerSample.container_barcode === container.container_barcode)
+            if (barcode_exists) {
+                barcodeExistsError()
+            }
+            /* Check if the name is used in the destinationContainerList */
+            let name_exists: boolean = destinationContainerList.some(containerSample => containerSample.container_name === container.container_name)
+            if (name_exists) {
+                nameExistsError()
+            }
+            exists = name_exists || barcode_exists
+            if (exists) {
+                /* Fast return if already exists in frontend display */
+                return exists
+            }
+    
+            /* Get existing containers from freezeman */
+            /* Check if there is already a container in Freezeman with that abrcode */
+            const barcodeResult = await dispatch(api.containers.list({ barcode: container.container_barcode }))
+            barcode_exists = barcodeResult.data.count != 0
+            if (barcode_exists) {
+                barcodeExistsError()
+            }
+            /* Check if there is already a container in Freezeman with that name */
+            const nameResult = await dispatch(api.containers.list({ name: container.container_name }))
+            name_exists = nameResult.data.count != 0
+            if (name_exists) {
+                nameExistsError()
+            }
+            exists = name_exists || barcode_exists
+            return exists
+        }    
+
+        const addContainer = (container: DestinationContainer) => {
+            onConfirm(container)
+            setIsPopup(false)
+            setNewContainer({ samples: {} })
         }
         const container = selectedTab == "load" ? loadedContainer : newContainer
         container.container_name = container.container_name ? container.container_name : container.container_barcode
         if (container.container_barcode && container.container_kind) {
-          if (selectedTab == "load") {
-            addContainer(container)
-          }
-          else {
-            containerAlreadyExists(container, destinationContainerList).then(exists => {
-              if (!exists) {
-                const barCodeResults = barcodeRules.filter((rule) => !(rule.pattern as RegExp).test(container.container_barcode))
-                const nameResults = nameRules.filter((rule) => !(rule.pattern as RegExp).test(container.container_name))
-                if (barCodeResults.length > 0) {
-                  setError("Invalid new container")
-                  const INVALID_BARCODE_NOTIFICATION_KEY = `LabworkStep.placement-invalid-container-barcode`
-                  notification.error({
-                    message: `Container Barcode -- ${barCodeResults.map((rule) => rule.message).join(" ")}`,
-                    key: INVALID_BARCODE_NOTIFICATION_KEY,
-                    duration: 20
-                  })
-                } if (nameResults.length > 0) {
-                  setError("Invalid new container")
-                  const INVALID_NAME_NOTIFICATION_KEY = `LabworkStep.placement-invalid-container-name`
-                  notification.error({
-                    message: `Container Name -- ${nameResults.map((rule) => rule.message).join(" ")}`,
-                    key: INVALID_NAME_NOTIFICATION_KEY,
-                    duration: 20
-                  })
-                } else {
-                  addContainer(container)
-                }
-               }
-            })
-          }
+            if (selectedTab == "load") {
+                addContainer(container as DestinationContainer)
+            }
+            else {
+                containerAlreadyExists(container as DestinationContainer, existingContainers).then(exists => {
+                    if (!exists) {
+                        const barCodeResults = barcodeRules.filter((rule) => !rule.pattern.test(container.container_barcode as string))
+                        const nameResults = nameRules.filter((rule) => !rule.pattern.test(container.container_name as string))
+                        if (barCodeResults.length > 0) {
+                            const INVALID_BARCODE_NOTIFICATION_KEY = `LabworkStep.placement-invalid-container-barcode`
+                            notification.error({
+                                message: `Container Barcode -- ${barCodeResults.map((rule) => rule.message).join(" ")}`,
+                                key: INVALID_BARCODE_NOTIFICATION_KEY,
+                                duration: 20
+                            })
+                        } if (nameResults.length > 0) {
+                            const INVALID_NAME_NOTIFICATION_KEY = `LabworkStep.placement-invalid-container-name`
+                            notification.error({
+                                message: `Container Name -- ${nameResults.map((rule) => rule.message).join(" ")}`,
+                                key: INVALID_NAME_NOTIFICATION_KEY,
+                                duration: 20
+                            })
+                        } else {
+                            addContainer(container as DestinationContainer)
+                        }
+                    }
+                })
+            }
         } else {
-            setError("Invalid container kind")
             const INVALID_KIND_NOTIFICATION_KEY = `LabworkStep.placement-invalid-container-kind`
             notification.error({
-              message: `Invalid destination container kind.`,
-              key: INVALID_KIND_NOTIFICATION_KEY,
-              duration: 20
+                message: `Invalid destination container kind.`,
+                key: INVALID_KIND_NOTIFICATION_KEY,
+                duration: 20
             })
         }
-    }, [loadedContainer, selectedTab, newContainer, destinationContainerList])
+    }, [selectedTab, loadedContainer, newContainer, dispatch, onConfirm, existingContainers])
 
-    const handleOnChange = useCallback((e, name) => {
-        const container = { ...newContainer }
+    const handleOnChange = useCallback((e: any, name: keyof typeof newContainer) => {
+        const container: typeof newContainer = { ...newContainer }
         container[name] = e.target ? e.target.value : e
         setNewContainer(container)
     }, [newContainer])
-
 
     return (
         <>
@@ -180,15 +180,15 @@ const AddPlacementContainer = ({ onConfirm, destinationContainerList, setDestina
                 <div style={{ display: 'flex', flexDirection: 'column' }}>
                     <Tabs defaultActiveKey={'New'} activeKey={selectedTab} onTabClick={(e) => setSelectedTab(e)}>
                         <Tabs.TabPane tab='New Container' key={'new'}>
-                          <Row style={{padding: "10px"}}>
-                            <Input value={newContainer.container_barcode} placeholder="Barcode" onChange={(e) => handleOnChange(e, 'container_barcode')} maxLength={MAX_CONTAINER_BARCODE_LENGTH}></Input>
-                          </Row>
-                          <Row style={{padding: "10px"}}>
-                            <Input value={newContainer.container_name} placeholder="Name (optional)" onChange={(e) => handleOnChange(e, 'container_name')} maxLength={MAX_CONTAINER_NAME_LENGTH}></Input>
-                          </Row>
-                          <Row style={{padding: "10px"}}>
-                            <Select value={newContainer.container_kind} clearIcon placeholder="Container kind" onChange={(e) => handleOnChange(e, 'container_kind')} style={{ width: "100%" }} options={getContainerKindOptions()}></Select>
-                          </Row>
+                            <Row style={{ padding: "10px" }}>
+                                <Input value={newContainer.container_barcode} placeholder="Barcode" onChange={(e) => handleOnChange(e, 'container_barcode')} maxLength={MAX_CONTAINER_BARCODE_LENGTH}></Input>
+                            </Row>
+                            <Row style={{ padding: "10px" }}>
+                                <Input value={newContainer.container_name} placeholder="Name (optional)" onChange={(e) => handleOnChange(e, 'container_name')} maxLength={MAX_CONTAINER_NAME_LENGTH}></Input>
+                            </Row>
+                            <Row style={{ padding: "10px" }}>
+                                <Select value={newContainer.container_kind} clearIcon placeholder="Container kind" onChange={(e) => handleOnChange(e, 'container_kind')} style={{ width: "100%" }} options={getContainerKindOptions()}></Select>
+                            </Row>
                         </Tabs.TabPane>
                         <Tabs.TabPane tab='Load Container' key={'load'}>
                             <SearchContainer exceptKinds={["tube"]} handleOnChange={(value) => handleContainerLoad(value)} />
